requirements.txt
matplotlib
"jax[cuda12_pip]" -f https://storage.googleapis.com/jax-releases/jax_cuda_releases.html
python-rtmidi
numpy
opencv-python
networkx
cupy-cuda12x
pyserial
